removed arrow from tooltip
